Tests/LibWeb/Text/expected/DOM/CharacterData-substringData-break-surrogate-pair.txt

Before substringData:
[0]: 55357
[1]: 56374
After substringData(0, 1):
[0]: 55357
